+++ src/main/java/ve/com/cge/appinvoice/accpay/dto/ShoppingDTO.java
@@ -31,7 +31,7 @@ import ve.com.cge.appinvoice.finance.model.Taxes;
  */
 public class ShoppingDTO {
  
-    private Integer id;
+    private Long id;
     private String description;
     private List<ShoppingDetails> shoppingDetails;
     private Supplier supplier;
@@ -45,7 +45,7 @@ public class ShoppingDTO {
     public ShoppingDTO() {
     }
 
-    public ShoppingDTO(Integer id, String description, List<ShoppingDetails> shoppingDetails, Supplier supplier, Bank bank, Taxes taxes, Timestamp date, double subTotal, double amountTax, double total) {
+    public ShoppingDTO(Long id, String description, List<ShoppingDetails> shoppingDetails, Supplier supplier, Bank bank, Taxes taxes, Timestamp date, double subTotal, double amountTax, double total) {
         this.id = id;
         this.description = description;
         this.shoppingDetails = shoppingDetails;
@@ -58,11 +58,11 @@ public class ShoppingDTO {
         this.total = total;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
