+++ src/main/java/ve/com/cge/appinvoice/accrec/dto/SellerDTO.java
@@ -24,22 +24,22 @@ package ve.com.cge.appinvoice.accrec.dto;
  */
 public class SellerDTO {
 
-    private Integer id;
+    private Long id;
     private String description;
      
     public SellerDTO() {
     }
 
-    public SellerDTO(Integer id, String description) {
+    public SellerDTO(Long id, String description) {
         this.id = id;
         this.description = description;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
